+++ CHANGELOG.md
@@ -29,6 +29,7 @@ The following is a curated list of changes in the Enact limestone module, newest
 - `limestone/Card` `background` for captions
 - `limestone/Popup` to restore focus properly when popup hides
 - `limestone/QuickGuidePanels` navigation buttons position
+- `limestone/TabLayout` to maintain the same scroller position when `Tabs` are collapsed or expanded
 - `limestone/VideoPlayer` size of the icons from the control buttons
 
 ## [1.2.0] - 2025-07-11